Move the ExternalJwtVerifierImpl to it's own package
