+++ blueocean-jwt/src/main/resources/io/jenkins/blueocean/auth/jwt/impl/external/ExternalJWTConfiguration/config.jelly
